........S. [ZBXNEXT-6406] more code style fixes and changelog

+++ src/zabbix_server/pinger/pinger.c
@@ -289,7 +289,7 @@ static int	parse_key_params(const char *key, const char *host_addr, icmpping_t *
 	{
 		if (NULL == host_addr || '\0' == *host_addr)
 		{
-			zbx_snprintf(error, max_error_len,
+			zbx_snprintf(error, (size_t)max_error_len,
 						"Ping item has to have target or host interface specified.");
 			goto out;
 		}
